Fix temperature float value store format
With float(2, 2) it store 0.99 value

+++ db/sql/create_tables.sql
@@ -2,7 +2,7 @@ USE 'meteo'
 
 CREATE TABLE METEO_DATA_SMART_HOUSE (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
-    temperature FLOAT(2, 2) NOT NULL,
+    temperature FLOAT(4, 2) NOT NULL,
     humidity TINYINT NOT NULL,
     create_date DATETIME NOT NULL
 );
